ai_tutor_backend/requirements.txt
# Core Flask
flask==3.0.0
flask-cors==4.0.0

# Database
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9    # PostgreSQL adapter (for production)

# Authentication
flask-jwt-extended==4.6.0
passlib[bcrypt]==1.7.4    # bcrypt is more secure than default

# Environment variables
python-dotenv==1.0.0

# AI Integration
openai==1.3.8             # OpenAI API (easier than transformers for MVP)
# transformers==4.35.2    # Uncomment if you want local HuggingFace models

# Payments
stripe==7.8.0

# HTTP requests & utilities
requests==2.31.0
gunicorn==21.2.0

# Development & Testing (optional but recommended)
pytest==7.4.3            # for unit testing
pytest-flask==1.3.0      # Flask testing utilities

# Data validation (highly recommended)
marshmallow==3.20.1       # for API input/output validation
flask-marshmallow==0.15.0

# Security enhancements
flask-limiter==3.5.0      # rate limiting (prevent API abuse)

# Monitoring (optional for later)
# sentry-sdk[flask]==1.38.0  # error tracking in production
